+++ src/CPU.hpp
@@ -52,7 +52,7 @@ private:
 
     std::random_device rd{};
     std::mt19937 m_rng_gen{rd()};
-    std::uniform_int_distribution<uint8_t> m_rng_dist{};
+    std::uniform_int_distribution<uint16_t> m_rng_dist{};
 
     uint8_t get_x() const { return (m_state.opcode & 0x0F00) >> 8; }
     uint8_t get_y() const { return (m_state.opcode & 0x00F0) >> 4; }
